% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrosbr-package.R
\docType{package}
\name{carrosbr-package}
\alias{carrosbr}
\alias{carrosbr-package}
\title{carrosbr: What the Package Does (Title Case)}
\description{
Um pacote para ver dados de carros no Brasil.
}
\keyword{internal}
